Now, the command "make-distribution.sh" works. Also was add a new option, "--example", in that script to copy  "example"  folder to "dist" folder.

+++ dev/make-distribution.sh
@@ -35,6 +35,7 @@ DISTDIR="$SPARK_HOME/dist"
 MAKE_TGZ=false
 MAKE_PIP=false
 MAKE_R=false
+MAKE_EXAMPLE=false
 NAME=none
 MVN="$SPARK_HOME/build/mvn"
 
@@ -60,6 +61,10 @@ while (( "$#" )); do
       ;;
     --r)
       MAKE_R=true
+      ;;
+    --example)
+      MAKE_EXAMPLE=true
+      echo "WARN: The operation '--example' is not yet supported"
       ;;
     --mvn)
       MVN="$2"
@@ -175,20 +180,25 @@ if [ -f "$SPARK_HOME"/common/network-yarn/target/scala*/spark-*-yarn-shuffle.jar
 fi
 
 # Copy examples and dependencies
-mkdir -p "$DISTDIR/examples/jars"
-cp "$SPARK_HOME"/examples/target/scala*/jars/* "$DISTDIR/examples/jars"
+if [ "$MAKE_EXAMPLE" == "true" ]; then
+    mkdir -p "$DISTDIR/examples/jars"
+    cp "$SPARK_HOME"/examples/target/scala*/jars/* "$DISTDIR/examples/jars"
 
-# Deduplicate jars that have already been packaged as part of the main Spark dependencies.
-for f in "$DISTDIR"/examples/jars/*; do
-  name=$(basename "$f")
-  if [ -f "$DISTDIR/jars/$name" ]; then
-    rm "$DISTDIR/examples/jars/$name"
-  fi
-done
+    # Deduplicate jars that have already been packaged as part of the main Spark dependencies.
+    for f in "$DISTDIR"/examples/jars/*; do
+      name=$(basename "$f")
+      if [ -f "$DISTDIR/jars/$name" ]; then
+        rm "$DISTDIR/examples/jars/$name"
+      fi
+    done
 
-# Copy example sources (needed for python and SQL)
-mkdir -p "$DISTDIR/examples/src/main"
-cp -r "$SPARK_HOME/examples/src/main" "$DISTDIR/examples/src/"
+    # Copy example sources (needed for python and SQL)
+    mkdir -p "$DISTDIR/examples/src/main"
+    cp -r "$SPARK_HOME/examples/src/main" "$DISTDIR/examples/src/"
+else
+    echo "Skipping building example distribution package"
+fi
+
 
 # Copy license and ASF files
 cp "$SPARK_HOME/LICENSE" "$DISTDIR"
